refactor(phase-2h): move execute_script_with_data to RailsRunnerService
Continues the openproject_client.py god-class decomposition from ADR-002.
The ~325-line ``execute_script_with_data`` method — the workhorse that
serialises Python data into Ruby, ships the script + data file into the
container, and parses the Rails-console JSON output — moves into
``OpenProjectRailsRunnerService`` alongside the rest of the Rails-execute
family. ``OpenProjectClient`` keeps a thin delegator so existing callers
(internal client methods, services, migrations) work unchanged.

Numbers
-------
- ``openproject_client.py``: 5,916 → 5,601 LOC (-315)
- ``openproject_rails_runner_service.py``: 558 → 877 LOC (+319)
- Cumulative across phases 2a–2h: ``openproject_client.py``
  7,342 → 5,601 LOC (-1,741, -23.7%)

Verification
------------
- pytest tests/unit: 953 passed (937 baseline + 16 provenance regression)
- mypy src/: clean (101 files)
- ruff check / ruff format: clean

No behavioural changes — pure mechanical extraction with self-references
rewritten to delegate through ``self._client`` and a small set of regex
constants imported lazily from the client module.
