+++ src/clients/openproject_client.py
@@ -5,7 +5,6 @@ import os
 import random
 import re
 import shlex
-import shutil
 import subprocess
 import time
 from collections.abc import Callable, Iterator
@@ -460,323 +459,9 @@ class OpenProjectClient:
     ) -> dict[str, Any]:
         """Execute a Ruby script in the Rails console with structured input data.
 
-        The data is serialized to JSON, transferred to the container under /tmp,
-        and made available to the Ruby script as the variable `input_data`.
-        The script is also written to a temporary file and executed via:
-        load "/tmp/<script>.rb"
-
-        The Ruby script should print a JSON payload between the markers
-        JSON_OUTPUT_START and JSON_OUTPUT_END, which will be parsed and returned
-        as `data` in the result dict.
-
-        Args:
-            script_content: Ruby code that expects `input_data` to be defined
-            data: Arbitrary JSON-serializable input for the script
-            timeout: Optional Rails console timeout
-
-        Returns:
-            Dict with keys: status ("success"|"error"), message, data (parsed JSON), output (raw snippet)
-
+        Thin delegator over ``self.rails_runner.execute_script_with_data``.
         """
-        # Prepare local temp paths
-        temp_dir = Path(self.file_manager.data_dir) / "temp_scripts"
-        temp_dir.mkdir(parents=True, exist_ok=True)
-
-        local_data_path = temp_dir / f"openproject_input_{os.urandom(4).hex()}.json"
-        try:
-            with local_data_path.open("w", encoding="utf-8") as f:
-                json.dump(data, f)
-        except Exception as e:
-            err_msg = f"Failed to serialize input data: {e}"
-            raise QueryExecutionError(err_msg) from e
-
-        # Compose Ruby script with a small header that loads JSON into `input_data`
-        container_data_path = Path("/tmp") / local_data_path.name
-        header = f"require 'json'\ninput_data = JSON.parse(File.read('{container_data_path.as_posix()}'))\n"
-        full_script = header + script_content
-
-        local_script_path: Path | None = None
-        container_script_path: Path | None = None
-        operation_succeeded = False  # Track success for debug file preservation
-        try:
-            # Create local script file and transfer both script and data
-            local_script_path = self._create_script_file(full_script)
-            container_script_path = self._transfer_rails_script(local_script_path)
-
-            # Transfer the input JSON to container
-            self.transfer_file_to_container(local_data_path, container_data_path)
-
-            # Execute the script inside Rails console.
-            # IMPORTANT: We use a unique execution ID to distinguish this run's output
-            # from any previous runs still visible in the tmux buffer.
-            exec_id = os.urandom(8).hex()
-            unique_start_marker = f"JSON_OUTPUT_START_{exec_id}"
-            unique_end_marker = f"JSON_OUTPUT_END_{exec_id}"
-
-            load_cmd = f'load "{container_script_path.as_posix()}"'
-            try:
-                target = self.rails_client._get_target()
-                tmux = shutil.which("tmux") or "tmux"
-
-                # First, send a Ruby command to define the unique markers for this execution
-                # The script will use these instead of hardcoded markers
-                marker_setup = f"$j2o_start_marker = '{unique_start_marker}'; $j2o_end_marker = '{unique_end_marker}'"
-                subprocess.run(
-                    [tmux, "send-keys", "-t", target, marker_setup, "Enter"],
-                    capture_output=True,
-                    text=True,
-                    check=True,
-                )
-                time.sleep(0.1)
-
-                # Send the load command
-                escaped_cmd = self.rails_client._escape_command(load_cmd)
-                subprocess.run(
-                    [tmux, "send-keys", "-t", target, escaped_cmd, "Enter"],
-                    capture_output=True,
-                    text=True,
-                    check=True,
-                )
-
-                # Poll for our unique JSON_OUTPUT_END marker with timeout
-                effective_timeout = timeout or self.rails_client.command_timeout
-                start_time = time.time()
-                output = ""
-                found_markers = False
-
-                while time.time() - start_time < effective_timeout:
-                    cap = subprocess.run(
-                        [tmux, "capture-pane", "-p", "-S", "-2000", "-t", target],
-                        capture_output=True,
-                        text=True,
-                        check=True,
-                    )
-                    output = cap.stdout
-
-                    # Normalize output by removing newlines to handle markers split
-                    # by terminal width wrapping in tmux pane capture
-                    normalized = output.replace("\n", "").replace("\r", "")
-
-                    # Check for markers in normalized output (handles line wrapping)
-                    # Use rfind to find LAST occurrence (actual JSON output, not command echo)
-                    # The command echo contains markers inside quotes which we must skip
-                    if unique_start_marker in normalized and unique_end_marker in normalized:
-                        # Find the LAST occurrence of start marker followed by JSON content
-                        start_pos = normalized.rfind(unique_start_marker)
-                        if start_pos != -1:
-                            # Verify this is actual JSON output (followed by [ or {), not command echo (followed by ')
-                            next_char_pos = start_pos + len(unique_start_marker)
-                            if next_char_pos < len(normalized):
-                                next_char = normalized[next_char_pos]
-                                if next_char in "[{":  # Actual JSON content
-                                    end_pos = normalized.find(unique_end_marker, next_char_pos)
-                                    if end_pos != -1 and end_pos > start_pos:
-                                        found_markers = True
-                                        output = normalized  # Use normalized for extraction
-                                        break
-
-                    time.sleep(0.2)  # Poll every 200ms
-
-                if not found_markers:
-                    logger.warning(
-                        "JSON_OUTPUT_END_%s marker not found within %d seconds",
-                        exec_id,
-                        effective_timeout,
-                    )
-            except Exception as e:
-                # Fallback: if Rails console crashed or is unstable (e.g., Reline/IRB errors),
-                # execute via non-interactive runner to avoid TTY/Reline issues.
-                from src.clients.rails_console_client import (
-                    CommandExecutionError,
-                    ConsoleNotReadyError,
-                    RubyError,
-                )
-
-                if isinstance(e, (ConsoleNotReadyError, CommandExecutionError, RubyError)):
-                    # Respect configuration: only fall back to rails runner when explicitly enabled
-                    if not config.migration_config.get("enable_runner_fallback", False):
-                        raise
-                    logger.warning(
-                        "Rails console execution failed (%s). Falling back to rails runner.",
-                        type(e).__name__,
-                    )
-                    # Try common app roots: /app (official) then /opt/openproject (legacy)
-                    runner_cmd = (
-                        f"(cd /app || cd /opt/openproject) && "
-                        f"bundle exec rails runner {container_script_path.as_posix()}"
-                    )
-                    # Ensure runner executes with the provided timeout (default to 120s if None)
-                    stdout, stderr, rc = self.docker_client.execute_command(
-                        runner_cmd,
-                        timeout=timeout or 120,
-                    )
-                    if rc != 0:
-                        q_msg = f"rails runner failed (rc={rc}): {stderr[:500]}"
-                        raise QueryExecutionError(q_msg) from e
-                    output = stdout
-                else:
-                    raise
-
-            # Extract JSON payload between unique markers (JSON_OUTPUT_START_{exec_id} / JSON_OUTPUT_END_{exec_id})
-            # Normalize output by removing newlines to handle markers split by terminal width wrapping
-            # This is safe because JSON from Rails is output on a single line
-            normalized_output = output.replace("\n", "").replace("\r", "")
-
-            # Find markers in normalized output - use rfind to get the LAST occurrence
-            # (the first occurrence is in the command echo: $j2o_start_marker = '...')
-            # The actual JSON output marker comes after the command echo
-            start_idx = normalized_output.rfind(unique_start_marker)
-            if start_idx != -1:
-                end_idx = normalized_output.find(unique_end_marker, start_idx + len(unique_start_marker))
-            else:
-                end_idx = -1
-
-            # Use normalized output for extraction since JSON has no real newlines
-            output = normalized_output
-
-            # DEBUG: Log marker positions
-            logger.debug(
-                "Marker extraction: start_idx=%s, end_idx=%s, marker=%s, output_len=%d",
-                start_idx,
-                end_idx,
-                exec_id,
-                len(output),
-            )
-            if start_idx != -1 and end_idx != -1:
-                json_preview = output[start_idx : end_idx + len(unique_end_marker)][:200]
-                logger.debug("JSON region preview: %r", json_preview)
-
-            if start_idx != -1 and end_idx != -1 and end_idx > start_idx:
-                json_str = output[start_idx + len(unique_start_marker) : end_idx].strip()
-
-                # Fallback parsing with sanitization to guard against stray control chars from IRB/tmux
-                def _try_parse(s: str) -> Any:
-                    return json.loads(s)
-
-                def _sanitize_control_chars(s: str) -> str:
-                    s = _RE_ANSI_ESCAPE.sub("", s)
-                    s = _RE_OSC_ESCAPE.sub("", s)
-                    s = _RE_OTHER_ESCAPE.sub("", s)
-                    return _RE_CTRL_CHARS.sub("", s)
-
-                def _extract_first_json_block(s: str) -> str | None:
-                    # Attempt to isolate the first balanced JSON object/array
-                    for i, ch in enumerate(s):
-                        if ch in "{[":
-                            opening = ch
-                            closing = "}" if ch == "{" else "]"
-                            depth = 0
-                            in_str = False
-                            esc = False
-                            for j in range(i, len(s)):
-                                c = s[j]
-                                if in_str:
-                                    if esc:
-                                        esc = False
-                                    elif c == "\\":
-                                        esc = True
-                                    elif c == '"':
-                                        in_str = False
-                                elif c == '"':
-                                    in_str = True
-                                elif c == opening:
-                                    depth += 1
-                                elif c == closing:
-                                    depth -= 1
-                                    if depth == 0:
-                                        return s[i : j + 1]
-                            break
-                    return None
-
-                # Always sanitize first - tmux output often has ANSI codes and line wraps
-                json_str = _sanitize_control_chars(json_str)
-                try:
-                    parsed = _try_parse(json_str)
-                except Exception:
-                    try:
-                        parsed = _try_parse(json_str)  # Already sanitized
-                    except Exception:
-                        candidate = _extract_first_json_block(json_str)
-                        if candidate is None:
-                            candidate = _extract_first_json_block(_sanitize_control_chars(json_str)) or json_str
-                        try:
-                            parsed = _try_parse(candidate)
-                        except json.JSONDecodeError as e:
-                            # Debug: log the problematic area
-                            pos = e.pos if hasattr(e, "pos") else 0
-                            start_ctx = max(0, pos - 20)
-                            end_ctx = min(len(candidate), pos + 20)
-                            ctx = candidate[start_ctx:end_ctx]
-                            char_at_pos = repr(candidate[pos : pos + 5]) if pos < len(candidate) else "EOF"
-                            logger.warning(
-                                "JSON parse error at pos %d: char=%s, context=%r",
-                                pos,
-                                char_at_pos,
-                                ctx,
-                            )
-                            q_msg = f"Failed to parse JSON output: {e}"
-                            raise QueryExecutionError(q_msg) from e
-                        except Exception as e:
-                            q_msg = f"Failed to parse JSON output: {e}"
-                            raise QueryExecutionError(q_msg) from e
-
-                # Return a structured success response
-                operation_succeeded = True
-                return {
-                    "status": "success",
-                    "message": "Script executed successfully",
-                    "data": parsed,
-                    "output": output[:2000],
-                }
-
-            # No JSON markers found - return an error envelope for callers
-            # Note: Still mark as "succeeded" since execution completed (just no JSON found)
-            operation_succeeded = True
-            return {
-                "status": "error",
-                "message": "JSON markers not found in Rails output",
-                "output": output[:2000],
-            }
-
-        finally:
-            # Cleanup logic: preserve debug files on errors if configured
-            preserve_on_error = config.migration_config.get("preserve_debug_files_on_error", True)
-            should_cleanup = operation_succeeded or not preserve_on_error
-
-            if not should_cleanup:
-                # Preserve files for debugging - log their locations
-                logger.warning(
-                    "Preserving debug files due to error (set preserve_debug_files_on_error=false to auto-cleanup):\n"
-                    "  Local script: %s\n"
-                    "  Container script: %s\n"
-                    "  Local data: %s\n"
-                    "  Container data: %s",
-                    local_script_path,
-                    container_script_path,
-                    local_data_path,
-                    container_data_path,
-                )
-            else:
-                # Best-effort cleanup of local + remote files
-                try:
-                    if local_script_path is not None and container_script_path is not None:
-                        self._cleanup_script_files(local_script_path, container_script_path)
-                except Exception as cleanup_err:
-                    logger.warning(
-                        "Failed to cleanup script files (local=%s, container=%s): %s",
-                        local_script_path,
-                        container_script_path,
-                        cleanup_err,
-                    )
-                try:
-                    self._cleanup_script_files(local_data_path, container_data_path)
-                except Exception as cleanup_err:
-                    logger.warning(
-                        "Failed to cleanup data files (local=%s, container=%s): %s",
-                        local_data_path,
-                        container_data_path,
-                        cleanup_err,
-                    )
+        return self.rails_runner.execute_script_with_data(script_content, data, timeout)
 
     def transfer_file_to_container(
         self,
